requirements.txt
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
pytz==2023.3